Moves Programme/CourseNav to signpost folder
