--- packages/lms/src/lib/components/navigation/IconNav.svelte
@@ -1,10 +1,0 @@
-<script lang="ts">
-  import { AppRail, AppRailTile } from '@skeletonlabs/skeleton';
-  import AddRemoveBookmarks from './icons/AddRemoveBookmarks.svelte';
-  import Bookmarks from './icons/Bookmarks.svelte';
-</script>
-
-<AppRail width="w-[3rem] sm:w-20">
-  <Bookmarks />
-  <AddRemoveBookmarks />
-</AppRail>
