+++ packages/lms/src/lib/components/signpost/CourseNav.svelte
